Added DependencyTree class.
This class exists as an alternative to the tree_libs function.  It is
designed to resolve dependencies recursively.  This works with the
improved resolve_rpath function to handle `@loader_path` on
dependencies.

The `tree_libs` function uses this class now but using its full
features seems to break everything using that function.

The Python warning is now replaced by `logger.error`.