Modify failing test.
delocate_tree_libs's documentation says that libraries not in root_path
are copied, but this test assumed that a dummy directory would not
cause libraries to be copied, or would modify libraries outside of
root_path and lib_path.

This test at the end of `test_delocate_tree_libs` seemed to use libraries
from a recent call to `_make_libtree` mixed with a `copied` variable from
an older call to `delocate_tree_libs`.  This test is now on its own to
prevent confusion.